pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-mcp-extended"
version = "2.1.0"
description = "Extended MCP server for Obsidian with plugin control (Tasks, Dataview, Kanban, Canvas), backlinks, tag management, and smart insertion - hybrid filesystem/API architecture"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Nate Strong", email = "nate@example.com"},
    {name = "Aleksa", email = "aleksa@example.com"}
]
keywords = ["mcp", "obsidian", "fastmcp", "llm", "ai", "notes", "knowledge-management", "backlinks", "tags", "statistics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastmcp>=0.5.8",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "urllib3>=2.0.0",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

[project.urls]
Homepage = "https://github.com/natestrong/obsidian-mcp"
Repository = "https://github.com/natestrong/obsidian-mcp"
Issues = "https://github.com/natestrong/obsidian-mcp/issues"

[project.scripts]
obsidian-mcp = "src.server:main"

[tool.setuptools]
packages = ["src", "src.tools", "src.utils", "src.models"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["py.typed"]